llm-entity-resolution/requirements.txt
openai>=1.12.0
dspy>=2.6.0
litellm>=1.30.0
langfuse>=2.40.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
pytest>=7.4.0
